gpu_service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
numpy>=1.26.0
soundfile>=0.12.0
librosa>=0.10.0
torch>=2.0.0
pyannote.audio>=3.3.0
speechbrain>=1.0.0
funasr>=1.3.0
python-multipart>=0.0.9
pydub>=0.25.0
faiss-cpu>=1.7.0
